separate objectid mongo of user value
