+++ backend/src/app/domain/values/User.ts
@@ -1,28 +1,8 @@
 import bcrypt from 'bcryptjs'
-import { isValidObjectId } from 'mongoose'
 
 import { IValueObject } from '@core/values/IValueObject'
 
 export type RoleType = 'admin' | 'disabled' | 'user'
-
-export class ObjectID implements IValueObject {
-  private _value: string
-
-  constructor(value: string) {
-    if (!isValidObjectId(value)) {
-      throw new Error(`The id ${value} is invalid objectId Mongo.`)
-    }
-    this._value = value
-  }
-
-  public toString(): string {
-    return this._value
-  }
-
-  public static toObjectID(value: string) {
-    return new ObjectID(value)
-  }
-}
 
 export class Name implements IValueObject {
   private _value: string
